pyproject.toml
[project]
name = "todoist-mcp-server"
version = "0.1.0"
description = "Todoist MCP Server for Claude Desktop - Manage your tasks with AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.9.4",
    "requests>=2.32.4",
    "python-dotenv>=1.1.0",
    "todoist-api-python>=3.1.0",
]
